add test and refactor id type

+++ platform/src/main/java/com/eventplatform/model/Event.java
@@ -1,10 +1,9 @@
 package com.eventplatform.model;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 public class Event extends Entity {
-    public Event(BigInteger id, Date creationDate) {
+    public Event(int id, Date creationDate) {
         super(id, creationDate);
     }
 }
